requirement.txt
mega.py: This library provides a simple and convenient interface to interact with the Mega cloud storage service. It allows you to easily upload, download, and manage files and folders in your Mega account.
python-telegram-bot: This library is a popular Python wrapper for the Telegram Bot API. It provides an easy-to-use interface for building Telegram bots that can send and receive messages, respond to user inputs, and perform other tasks.
urllib3: This is a powerful HTTP client library for Python that provides a wide range of features for working with HTTP requests and responses. It is used by many popular Python libraries and frameworks and provides a convenient interface for working with web APIs and other HTTP-based services.
requests: This is another popular HTTP client library for Python that provides a simple and easy-to-use interface for making HTTP requests. It is widely used in many Python applications and provides a range of features for working with HTTP requests and responses.
pillow: This library is a fork of the Python Imaging Library (PIL) that provides a range of features for working with images in Python. It allows you to easily read, write, and manipulate images, and provides a range of tools for working with different image formats.
Overall, these libraries provide a range of useful tools and features for building Python applications that interact with web services, manipulate images, and perform other tasks. By including these libraries in your requirements.txt file, you can ensure that your application has access to all the necessary dependencies to run smoothly and perform the tasks you need it to do.